Obstacles done

+++ Obstacle.h
@@ -5,7 +5,7 @@ class Obstacle :public Moveable
 public:
 	Obstacle(Tmpl8::vec2, Collider col);
 	bool operator==(const Obstacle& ob) {
-		return ob.pos.x == pos.x && ob.pos.y == ob.pos.y;
+		return static_cast<int>(ob.pos.x) == static_cast<int>(pos.x) && static_cast<int>(ob.pos.y) == static_cast<int>(ob.pos.y);
 	}bool operator!=(const Obstacle& ob) {
 		return ob.pos.x != pos.x || ob.pos.y != ob.pos.y;
 	}
